Fix reboot NPE by initializing FileUtils to ensure mounts are read
The main activity is calling RebootHelper.showRebootDialog which
calls reboot. This method uses RecoveryHelper.getRecoveryFilePath
which uses FileUtils.getPrimarySdCard. This method returns
a cached variable that should be initialized by the private method
readMounts called by FileUtils.init.

However, this init function was never called causing the cached
variable to be unitialized, hence resulting in a NPE. We fix
this problem by calling init in the main activity.

Change-Id: I19a921c45853a6ab5ba29d1f639d82a50ba99688