ParanoidHub: Unify download and installation process for A/B
The process change is to handle download and installation together when the user agrees to start the update. Since the A/B process is done in the background, there's no reason to present extra steps for installation. Once the download and install completes, just notify the user to restart the device. Or if the update fails in some way notify the user as well

Change-Id: Iacc4eaeea0163a884fed72f63f96b6d3f403f281