ParanoidHub: Add support for staged rollouts
This allows staged rollouts to reduce server load by making the update available to specific IMEIs after a certain period of time. All of this is handled on device, so no IMEI information is sent or stored to the server for privacy reasons.

This commit also adds support for device whitelisting by full imei, RolloutContractor retreives whitelisted info from the server and compares it with the device, if theres a match, it will bypass the rollout restrictions

Change-Id: I78245e8d62157515386f48c354aa0c08c6c4714d