ParanoidHub: Compare build type before getting current timestamp
Since our versioning has changed in Sapphire, we need to increase the number of hyphens when searching for the timestamp from our version prop. Non-release builds have 4 hyphens as they include the build type while a release build does not because it requires 3.

Also, remove the TYPE_DEV build type as it's no longer used in our versioning system

Change-Id: I3d5104b89c605cb9bca3b738da1fcc36b89919b5