ParanoidHub: Show the circle progress when launching OTA
This changes the logic behind the circle progress and the system searching for updates. By default the system checks the sever and downloads any new config files when launching the app. This would cause a delay when updating the UI, because the circle progress was not configured to show in this case but only for user initiated checks.

To avoid this, follow stock OEM behavior and show the circle progress upon launching OTA. This allows the UI to update properly after initial checks.

Also shorten the delay for non user initiated checks, so the user is not waiting 5 seconds or more for the UI to update upon launch

Change-Id: Idf43a58b9e7c1c9400b6946d9bbdc3abbe44a5de