ParanoidHub: Implement enrollment for Tester's Program
This replaces the old beta opt-in implementation toggle with a full fledged UI, rebranded to "Testers Program". The rebrand allows for users to opt-in to both Alpha's and Beta's at once and opt out. Opting out should provide the user with the latest public release, while opting in should provide the latest alpha or beta available for the device

- Also removed old preference menu

Change-Id: Ife06d899778214b9e0314ba84e3762b64bc5ad4d