ParanoidHub: Update to Hub 5.0
Hub has been rewritten completely from scratch with the goal to introduce a lighter frontend and backend system.

The results of the rewrite has reduced code duplication by 50% and created a solid foundation for the UI. The UI states and tasks are all broken up into individual components. A component stores the relevant information for each state and publishes the appropriate state. This on it's own keeps the UI from being in a buggy state for most usage cases.

Components and States are instances of a single interface which helps shrink code duplication as well.

Change-Id: Ie3f13cb5a58591e2e683c02c10b5ce99db3627c7