ParanoidHub: Add support for device changelogs
This commit adds support for device changelogs only generated and shown to the user based on the type of update.

Currently the new logic being applied is, if an incremental update is being released for said device, we compare the version to the main changelog which will always be higher based on our release cycle. Once that's confirms it we show the device changelog and any relevant changes instead of the main one.

ex.
Topaz 1 Beta (main release)
Topaz 1.1 Beta (device incremental)

In this case device changelog is shown since version is greater than main. If versions are the same, then we prioritize main changelog as that's assumed to be a global release

Change-Id: I4338518035646d84f3ad2bc311b04f447f923695